+++ custom_components/openpublictransport/providers/__init__.py
@@ -5,29 +5,53 @@ from typing import Dict, Optional, Type
 from homeassistant.core import HomeAssistant
 
 from ..const import (
+    PROVIDER_AVV_AUGSBURG,
+    PROVIDER_BEG,
+    PROVIDER_BSVG,
     PROVIDER_BVG,
+    PROVIDER_DING,
     PROVIDER_HVV,
     PROVIDER_KVV,
     PROVIDER_MVV,
     PROVIDER_NTA_IE,
+    PROVIDER_NVBW,
+    PROVIDER_NWL,
+    PROVIDER_OEBB,
     PROVIDER_RMV,
+    PROVIDER_RVV,
+    PROVIDER_SBB,
     PROVIDER_TRAFIKLAB_SE,
+    PROVIDER_TRANSITOUS,
     PROVIDER_VAGFR,
     PROVIDER_VGN,
+    PROVIDER_VRN,
     PROVIDER_VRR,
+    PROVIDER_VVO,
     PROVIDER_VVS,
 )
+from .avv import AVVProvider
 from .base import BaseProvider
+from .beg import BEGProvider
+from .bsvg import BSVGProvider
 from .bvg import BVGProvider
+from .ding import DINGProvider
 from .hvv import HVVProvider
 from .kvv import KVVProvider
 from .mvv import MVVProvider
 from .nta import NTAProvider
+from .nvbw import NVBWProvider
+from .nwl import NWLProvider
+from .oebb import OeBBProvider
 from .rmv import RMVProvider
+from .rvv import RVVProvider
+from .sbb import SBBProvider
 from .trafiklab import TrafiklabProvider
+from .transitous import TransitousProvider
 from .vagfr import VAGFRProvider
 from .vgn import VGNProvider
+from .vrn import VRNProvider
 from .vrr import VRRProvider
+from .vvo import VVOProvider
 from .vvs import VVSProvider
 
 _PROVIDER_REGISTRY: Dict[str, Type[BaseProvider]] = {}
@@ -70,3 +94,15 @@ register_provider(PROVIDER_VAGFR, VAGFRProvider)
 register_provider(PROVIDER_RMV, RMVProvider)
 register_provider(PROVIDER_TRAFIKLAB_SE, TrafiklabProvider)
 register_provider(PROVIDER_NTA_IE, NTAProvider)
+register_provider(PROVIDER_VRN, VRNProvider)
+register_provider(PROVIDER_VVO, VVOProvider)
+register_provider(PROVIDER_DING, DINGProvider)
+register_provider(PROVIDER_AVV_AUGSBURG, AVVProvider)
+register_provider(PROVIDER_RVV, RVVProvider)
+register_provider(PROVIDER_BSVG, BSVGProvider)
+register_provider(PROVIDER_NWL, NWLProvider)
+register_provider(PROVIDER_NVBW, NVBWProvider)
+register_provider(PROVIDER_BEG, BEGProvider)
+register_provider(PROVIDER_SBB, SBBProvider)
+register_provider(PROVIDER_OEBB, OeBBProvider)
+register_provider(PROVIDER_TRANSITOUS, TransitousProvider)